update test for get and delete, fix

+++ srcs/HttpResponse/DELETE/delete_target.cpp
@@ -5,9 +5,16 @@
 #include "HttpResponse.hpp"
 
 StatusCode HttpResponse::delete_target() {
-    std::string target_path = HttpResponse::get_resource_path();
-    DEBUG_PRINT(YELLOW, " DELETE path: %s", target_path.c_str());
+    Result<std::string, StatusCode> indexed_result = Config::get_indexed_path(this->server_config_,
+                                                                              this->request_.request_target());
+    if (indexed_result.is_err()) {
+        return indexed_result.get_err_value();
+    }
+    std::string path = indexed_result.get_ok_value();
+    DEBUG_PRINT(YELLOW, " DELETE path: %s", path.c_str());
 
-    FileHandler file(target_path);
-    return file.delete_file();
+    FileHandler file(path);
+    StatusCode result = file.delete_file();
+    // std::cout << CYAN << "  delete result: " << result << RESET << std::endl;
+    return result;
 }
